[app] wc3converter should support more formats
When USE_QBLP is specified wc3converter supports more image formats-
This can be used to convert BLP files into JPEG or PNG files.
Besides JPEG or PNG files can be converted back to BLP files.

Added a test case for the conversion.

+++ src/blp/jpeg.cpp
@@ -346,7 +346,7 @@ void readMipMapJpeg(Blp::MipMap &mipMap, byte *buffer, dword bufferSize)
 
 	if (!cinfo.saw_JFIF_marker)
 	{
-		std::cerr << boost::format(_("Warning: Did not find JFIF marker. JFIF format is used by default!\nThis is the JFIF version of the image %1%.%2%")) % cinfo.JFIF_major_version % cinfo.JFIF_minor_version << std::endl;
+		std::cerr << boost::format(_("Warning: Did not find JFIF marker. JFIF format is used by default!\nThis is the JFIF version of the image %1%.%2%")) % (int)cinfo.JFIF_major_version % (int)cinfo.JFIF_minor_version << std::endl;
 	}
 
 	jpeg_destroy_decompress(&cinfo);
